feat: Saving the configuration in local storage
